Fix CodingStyle except for the missing JavaDoc
