--- src/main/java/Fish/FishException.java
@@ -1,12 +1,0 @@
-package Fish;
-
-public class FishException extends Exception {
-    public FishException(String message) {
-        super(message);
-    }
-
-    @Override
-    public String getMessage() {
-        return super.getMessage();
-    }
-}
