--- src/main/java/Fish/Launcher.java
@@ -1,13 +1,0 @@
-package Fish;
-
-import com.sun.tools.javac.Main;
-import javafx.application.Application;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Fish.Main.class, args);
-    }
-}
