--- src/main/java/Fish/MainWindow.java
@@ -1,52 +1,0 @@
-package Fish;
-import Fish.ui.Fish;
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.VBox;
-/**
- * Controller for the main GUI.
- */
-public class MainWindow extends AnchorPane {
-    @FXML
-    private ScrollPane scrollPane;
-    @FXML
-    private VBox dialogContainer;
-    @FXML
-    private TextField userInput;
-    @FXML
-    private Button sendButton;
-
-    private Fish fish;
-
-    private Image userImage = new Image(this.getClass().getResourceAsStream("/xhs_azgc.jpeg"));
-    private Image fishImage = new Image(this.getClass().getResourceAsStream("/xhs_blush.jpeg"));
-
-    @FXML
-    public void initialize() {
-        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
-    }
-
-    /** Injects the Duke instance */
-    public void setFish(Fish f) {
-        fish = f;
-    }
-
-    /**
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
-     */
-    @FXML
-    private void handleUserInput() {
-        String input = userInput.getText();
-        String response = fish.getResponse(input);
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(input, userImage),
-                DialogBox.getDukeDialog(response, fishImage)
-        );
-        userInput.clear();
-    }
-}
