--- src/main/java/Fish/task/TaskList.java
@@ -1,36 +1,0 @@
-package Fish.task;
-
-import java.util.ArrayList;
-
-public class TaskList {
-
-    private ArrayList<Task> tasks;
-
-    public TaskList(ArrayList<Task> tasks) {
-        this.tasks = tasks;
-    }
-
-    public TaskList() {
-        this.tasks = new ArrayList<>();
-    }
-
-    public ArrayList<Task> getTasks() {
-        return tasks;
-    }
-
-    public void add(Task t) {
-        tasks.add(t);
-    }
-
-    public Task delete(int i) {
-        return tasks.remove(i);
-    }
-
-    public Task get(int i) {
-        return tasks.get(i);
-    }
-
-    public int size() {
-        return tasks.size();
-    }
-}
