+++ src/main/java/fish/FishException.java
@@ -1,0 +1,12 @@
+package fish;
+
+public class FishException extends Exception {
+    public FishException(String message) {
+        super(message);
+    }
+
+    @Override
+    public String getMessage() {
+        return super.getMessage();
+    }
+}
