+++ src/main/java/fish/Launcher.java
@@ -1,0 +1,12 @@
+package fish;
+
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(fish.Main.class, args);
+    }
+}
