+++ src/main/java/fish/command/DeleteCommand.java
@@ -1,0 +1,21 @@
+package fish.command;
+
+import fish.storage.Storage;
+import fish.task.TaskList;
+import fish.ui.Ui;
+
+public class DeleteCommand extends Command {
+    private final int idx;
+
+    public DeleteCommand(int idx) {
+        this.idx = idx;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        tasks.delete(this.idx);
+        int index = this.idx + 1;
+        ui.printIn("Booo t'as supprime le task " + index);
+        storage.save(tasks.getTasks());
+    }
+}
