+++ src/main/java/fish/command/ExitCommand.java
@@ -1,0 +1,18 @@
+package fish.command;
+
+import fish.storage.Storage;
+import fish.task.TaskList;
+import fish.ui.Ui;
+
+public class ExitCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.showExit();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
