+++ src/main/java/fish/task/Task.java
@@ -1,0 +1,49 @@
+package fish.task;
+
+public abstract class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void markAsUndone() {
+        this.isDone = false;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public boolean isDone() {
+        return isDone;
+    }
+
+    public void setIsDone(boolean isDone) {
+        this.isDone = isDone;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public abstract String getType();
+
+    public String toString() {
+        return "[" + getStatusIcon() + "] " + description;
+    }
+
+    public String toFileString() {
+        return String.join(" | ", getType(), isDone ? "1" : "0", description);
+    }
+}
